add ownership request handle page and server actions
