+++ packages/app/src/app/(default)/rattachement/page.tsx
@@ -1,3 +1,7 @@
+import { authConfig } from "@api/core-domain/infra/auth/config";
+import { redirect } from "next/navigation";
+import { getServerSession } from "next-auth";
+
 import { AddDeclarer } from "./AddDeclarer";
 
 const title = "Connexion";
@@ -9,10 +13,13 @@ export const metadata = {
   },
 };
 
-const RattachementPage = () => {
+const RattachementPage = async () => {
+  const session = await getServerSession(authConfig);
+  if (!session) redirect("/login");
+
   return (
     <>
-      <AddDeclarer />
+      <AddDeclarer email={session.user.email} />
     </>
   );
 };
