+++ packages/app/src/app/admin/rattachements/Pagination.tsx
@@ -1,23 +1,49 @@
 "use client";
 
+import Button from "@codegouvfr/react-dsfr/Button";
+import {
+  type GetOwnershipRequestDTO,
+  type GetOwnershipRequestInputSchemaDTO,
+} from "@common/core-domain/dtos/OwnershipRequestDTO";
 import { useEffect } from "react";
 
-import { ButtonGroup } from "../design-system/base/ButtonGroup";
-import { FormButton } from "../design-system/base/FormButton";
-import { useListeDeclarants } from "./useListeDeclarants";
-import { useOwnershipRequestListStore } from "./useOwnershipRequestListStore";
+import { type OwnershipRequestListStoreType, useOwnershipRequestListStore } from "./useOwnershipRequestListStore";
 
 type Props = {
   className?: string;
+  fetchedItems?: GetOwnershipRequestDTO;
 };
 
-export const Pagination = ({ className }: Props) => {
+const ITEMS_PER_PAGE = 10;
+
+export type SearchParams = Partial<Omit<OwnershipRequestListStoreType["formState"], "checkedItems" | "globalCheck">>;
+
+export const buildKey = (search?: SearchParams) => {
+  if (!search) return null;
+
+  const { orderBy, orderDirection, query, status, pageNumber, pageSize } = search;
+
+  const limit = pageSize || ITEMS_PER_PAGE;
+  const offset = pageNumber ? pageNumber * limit : 0;
+
+  const params = new URLSearchParams({
+    ...(search.orderBy && { orderBy }),
+    ...(search.orderDirection && { orderDirection }),
+    ...(search.query && { query }),
+    ...(search.status && { status }),
+    limit: String(limit),
+    offset: String(offset),
+  } as GetOwnershipRequestInputSchemaDTO);
+
+  return `${params.toString()}`;
+};
+
+export const Pagination = ({ className, fetchedItems }: Props) => {
   const formState = useOwnershipRequestListStore(state => state.formState);
   const firstPage = useOwnershipRequestListStore(state => state.firstPage);
   const nextPage = useOwnershipRequestListStore(state => state.nextPage);
   const previousPage = useOwnershipRequestListStore(state => state.previousPage);
 
-  const { fetchedItems } = useListeDeclarants(formState);
   const { pageSize, pageNumber } = formState;
 
   const totalCount = fetchedItems?.totalCount || 0;
@@ -35,7 +61,7 @@ export const Pagination = ({ className }: Props) => {
     if (fetchedItems && fetchedItems.data.length === 0 && formState.pageNumber > 0) {
       firstPage();
     }
-  }, [fetchedItems, firstPage, formState.pageNumber]);
+  }, [fetchedItems, firstPage, formState]);
 
   return (
     <nav aria-label="Pagination" className={className}>
@@ -53,14 +79,12 @@ export const Pagination = ({ className }: Props) => {
         </p>
       </div>
       <div>
-        <ButtonGroup inline="mobile-up" className="fr-mb-4w">
-          <FormButton disabled={!canGoToPreviousPage} onClick={previousPage}>
-            Précédent
-          </FormButton>
-          <FormButton disabled={!canGoToNextPage} variant="tertiary" onClick={nextPage}>
-            Suivant
-          </FormButton>
-        </ButtonGroup>
+        <Button disabled={!canGoToPreviousPage} onClick={previousPage}>
+          Précédent
+        </Button>
+        <Button priority="tertiary" disabled={!canGoToNextPage} onClick={nextPage}>
+          Suivant
+        </Button>
       </div>
     </nav>
   );
